- Now using glbinding for OpenGL functions, as it is more up-to-date with new OpenGL versions, source will be included  - Added empty class for SDL joysticks (GLFW's joysticks will be tricky in terms of mapping)

+++ opengl/components/coffeetexture.cpp
@@ -15,7 +15,7 @@ void CoffeeTexture::loadTexture()
     if(texture.isNull())
         return;
     texture = imageProcessor(texture);
-    textureHandle = TextureHelper::allocTexture(GL_RGBA,GL_RGBA,texture.width(),texture.height(),texture.bits());
+    textureHandle = TextureHelper::allocTexture(static_cast<int>(GL_RGBA),GL_RGBA,texture.width(),texture.height(),texture.bits());
     if(textureHandle>0)
         validTexture = true;
 }
